requirements.txt
PyExifTool >= 0.5.6, < 6
arrow >= 1.3.0, < 2
ffmpeg >= 1.4, < 2
